+++ examples/react-app/src/main.tsx
@@ -2,15 +2,17 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import App from "./App.tsx";
 import "./index.css";
-import { setup } from "./dojo/setup";
+import { setup } from "./dojo/generated/setup.ts";
 import { DojoProvider } from "./DojoContext";
+import { dojoConfig } from "./dojo/dojoConfig.ts";
 
 async function init() {
     const rootElement = document.getElementById("root");
     if (!rootElement) throw new Error("React root not found");
     const root = ReactDOM.createRoot(rootElement as HTMLElement);
 
-    const setupResult = await setup();
+    const setupResult = await setup(dojoConfig());
+
     root.render(
         <React.StrictMode>
             <DojoProvider value={setupResult}>
